+++ sdk/ai/Azure.AI.Projects/tests/AgentsTestBase.cs
@@ -56,6 +56,7 @@ public class AgentsTestBase : ProjectsClientTestBase
         ConnectedAgent,
         DeepResearch,
         AzureFunctionTool,
+        WorkIQTool,
     }
 
     public Dictionary<ToolType, string> ToolPrompts = new()
@@ -97,6 +98,7 @@ public class AgentsTestBase : ProjectsClientTestBase
         {ToolType.A2A, "What can the secondary agent do?"},
         {ToolType.A2ASpecialConnection, "What can the secondary agent do?"},
         {ToolType.AzureFunctionTool, "What is the most prevalent element in the universe? What would foo say?"},
+        {ToolType.WorkIQTool, "What meetings do I have scheduled today?"},
     };
 
     public Dictionary<ToolType, string> ToolInstructions = new()
@@ -131,6 +133,7 @@ public class AgentsTestBase : ProjectsClientTestBase
         {ToolType.MCPToolbox, "You are a helpful assistant." },
         {ToolType.A2A, "You are a helpful assistant."},
         {ToolType.A2ASpecialConnection, "You are a helpful assistant."},
+        {ToolType.WorkIQTool, "You are a helpful assistant that can access Microsoft 365 data through WorkIQ. Use the WorkIQ tool to search and retrieve information from emails, calendar events, Teams messages, and other Microsoft 365 content to assist users with their questions." },
     };
 
     public Dictionary<ToolType, string> ExpectedOutput = new()
@@ -199,6 +202,7 @@ public class AgentsTestBase : ProjectsClientTestBase
         {ToolType.FabricIQ, "mcp_call"},
         {ToolType.A2A, "a2a_preview_call_output"},
         {ToolType.A2ASpecialConnection, "a2a_preview_call_output"},
+        {ToolType.WorkIQTool, "a2a_preview_call_output"}
     };
     #endregion
 
@@ -597,6 +601,7 @@ public class AgentsTestBase : ProjectsClientTestBase
             ToolType.A2ASpecialConnection => GetA2ATool(true),
             ToolType.AzureFunction => GetFunctionTool(),
             ToolType.MCPToolbox => await GetToolBoxAsync(projectClient),
+            ToolType.WorkIQTool => new WorkIQPreviewTool(TestEnvironment.WORKIQ_CONNECTION_ID),
             _ => throw new InvalidOperationException($"Unknown tool type {toolType}")
         };
         string instructions;
@@ -681,7 +686,11 @@ public class AgentsTestBase : ProjectsClientTestBase
         List<string> hostedAgents = await projectClient.AgentAdministrationClient.GetAgentsAsync().Select((x) => x.Name).Where((x) => x.StartsWith(HOSTED_AGENT)).ToListAsync();
         foreach (string hostedAgent in hostedAgents)
         {
-            projectClient.AgentAdministrationClient.DeleteAgent(hostedAgent);
+            try
+            {
+                projectClient.AgentAdministrationClient.DeleteAgent(hostedAgent);
+            }
+            catch { }
         }
         await RemoveToolBoxMayBe(projectClient);
     }